Tweaking, proper overlays and interaction

+++ code/game/machinery/floor_shocker.dm
@@ -1,87 +1,98 @@
 /obj/item/weapon/floor_shocker
-	var/power_source
-	var/got_cell = FALSE
+	var/datum/powernet/powernet
+	var/obj/item/weapon/cell/large/cell
 	var/obj/parent
-	var/deployed = FALSE
 	var/active = FALSE
 
-	var/passive_power_use = 50
+	var/passive_power_use = 200
+	var/fire_delay = 40
 	var/radius = 2
 
 	icon = 'icons/obj/floor_shocker.dmi'
 	icon_state = "sapper"
 	name = "floor mounted shocker"
-	desc = "A nasty sapper that electrifies floors around."
+	desc = "A nasty sapper that electrifies floors around. Uses L-size batteries, but has connectors for outer wiring."
 	layer = LOW_OBJ_LAYER
-	var/turf_overlay_image
 
 /obj/item/weapon/floor_shocker/New()
-	turf_overlay_image = image(icon, "overlay_turf")
 	..()
 
 /obj/item/weapon/floor_shocker/Process()
-	if(!anchored) return
 	if(!active) return
-	if(!power_source) return
+	if(!cell && !powernet) return
+
+	var/got_power
+	if(powernet)
+		got_power = powernet.draw_power(passive_power_use)
+	else
+		got_power = cell.use(passive_power_use)
+
+	if(got_power == 0)
+		active = FALSE
+		update_icon()
+		return
+	if(!anchored)
+		explosion(get_turf(src), 0,1,2)
+
 	for(var/mob/living/mob in range(src, radius))
 		fire(mob)
 
 /obj/item/weapon/floor_shocker/attackby(var/obj/item/I, var/mob/living/user, var/params)
+	if(!Adjacent(user))
+		return FALSE
 	if(fire(user, TRUE))
 		return FALSE
 	add_fingerprint(user)
 	if(istype(I, /obj/item/weapon/cell/large))
-		if(!got_cell && power_source)
-			to_chat(user, "You added cell to shocker.")
+		if(!cell)
+			to_chat(user, SPAN_NOTICE("You added cell to shocker."))
 			user.drop_item()
 			I.forceMove(src)
-			power_source = I
-			got_cell = TRUE
+			cell = I
 			update_icon()
 			return
 	var/tool_type = I.get_tool_type(user, list(QUALITY_SCREW_DRIVING, QUALITY_BOLT_TURNING), src)
 	switch(tool_type)
 		if(QUALITY_BOLT_TURNING)
+			if(!anchored)
+				to_chat(user, SPAN_WARNING("It's not deployed yet."))
+				return
 			if(active)
 				to_chat(user, SPAN_WARNING("Turn it off first!"))
 				return
 			var/turf/turf = get_turf(src)
 			var/obj/structure/cable/C = turf.get_cable_node()
-			if(!got_cell & !power_source)
+			if(!cell)
 				if(!C)
-					to_chat(user, "You need to set up proper wiring there.")
+					to_chat(user, SPAN_WARNING("You need to set up proper wiring there."))
 					return
-			if(I.use_tool(user, src, WORKTIME_NORMAL, tool_type, FAILCHANCE_NORMAL, required_stat = STAT_MEC))
-				to_chat(user, SPAN_NOTICE("You [anchored?"un":""]anchored [src] [anchored?"from":"to"] floor."))
-				anchored = !anchored
-				if(!got_cell)
-					if(!anchored)
-						power_source = null
-					else
-						power_source = C.powernet
+			if(I.use_tool(user, src, WORKTIME_FAST, tool_type, FAILCHANCE_NORMAL, required_stat = STAT_MEC))
+				to_chat(user, SPAN_NOTICE("You unanchored [src] from floor."))
+				anchored = FALSE
+				powernet = null
 				update_icon()
 			return
 		if(QUALITY_SCREW_DRIVING)
 			if(active)
 				to_chat(user, SPAN_WARNING("Turn it off first!"))
 				return
-			if(!got_cell)
+			if(!cell)
 				to_chat(user, SPAN_WARNING("There is nothing to do!"))
 				return
-			if(I.use_tool(user, src, WORKTIME_NORMAL, tool_type, FAILCHANCE_NORMAL, required_stat = STAT_MEC))
+			if(I.use_tool(user, src, WORKTIME_FAST, tool_type, FAILCHANCE_NORMAL, required_stat = STAT_MEC))
 				to_chat(user, SPAN_NOTICE("You detached cell from [src]"))
-				var/obj/item/cell = power_source
 				cell.forceMove(get_turf(user))
-				power_source = null
+				cell = null
 				update_icon()
 			return
 	return ..()
 /obj/item/weapon/floor_shocker/attack_self(var/mob/living/carbon/human/user)
-	if(anchored || deployed)//how the fuck.
+	if(anchored)//how the fuck.
 		CRASH()
-		return FALSE
+
 	var/turf/turf = get_step(get_turf(user), user.dir)
 	var/obj/structure/table/table = locate(/obj/structure/table/) in turf
+
 	if(table)
 		if(table.material)
 			if(table.material == MATERIAL_STEEL \
@@ -91,64 +102,79 @@
 			|| table.material == MATERIAL_PLASTEEL \
 			|| table.material == MATERIAL_TITANIUM \
 			|| table.material == MATERIAL_PLATINUM) //urgh
-				if(deploy(user, turf, table))
-					return TRUE
-	var/obj/machinery/machine = locate(/obj/machinery) in turf
+				return deploy(user, turf, table)
+	var/obj/machinery/power/machine = locate(/obj/machinery/power) in turf
 	if(machine)
-		if(deploy(user, turf, machine))
-			return TRUE
+		return deploy(user, turf, machine)
+	to_chat(user, SPAN_WARNING("You need something conductive to attach to!"))
 	return FALSE
 
 /obj/item/weapon/floor_shocker/proc/deploy(var/mob/living/carbon/human/user, var/turf/turf, var/obj/parent)
-	if(!power_source)
-		var/obj/structure/cable/C = turf.get_cable_node()
-		if(!C)
-			to_chat(user, "You need to set up proper wiring there.")
-			return FALSE
-		power_source = C.powernet
-	if(fire(user, TRUE))
+	var/obj/structure/cable/C = turf.get_cable_node()
+	if(C)
+		powernet = C.powernet
+	else if(!cell)
+		to_chat(user, SPAN_WARNING("You need to set up proper wiring there."))
 		return FALSE
+
+	to_chat(user, SPAN_NOTICE("You started deploying [src]..."))
 	if(!do_after(user, 100))
 		return FALSE
+	to_chat(user, SPAN_NOTICE("You deployed [src]."))
 
 	src.parent = parent
 	user.drop_from_inventory(src)
 	src.forceMove(turf)
-	deployed = TRUE
+	anchored = TRUE
 	update_icon()
 	return TRUE
 
 /obj/item/weapon/floor_shocker/attack_hand(var/mob/living/carbon/human/user)
-	if(!deployed)
-		..()
-	if(anchored && power_source && !fire(user, TRUE))
+	if(!anchored)
+		return ..()
+	if(!Adjacent(user))
+		return FALSE
+	if(!cell && !powernet)
+		to_chat(user, "Device is unpowered.")
+		return FALSE
+	if(!fire(user, TRUE))
 		active = !active
 		update_icon()
 		var/datum/effect/effect/system/spark_spread/spark_system = new ()
 		spark_system.set_up(5, 0, src)
 		spark_system.attach(src)
 		spark_system.start()
-		spawn(10)
+		spawn(fire_delay)
 			qdel(spark_system)
-		return
+			while(active)
+				Process()
+				sleep(fire_delay)
+		return TRUE
 
 /obj/item/weapon/floor_shocker/update_icon()
 	overlays.Cut()
-	icon_state = "sapper[deployed?"_deployed":""][active?"_on":""]"
-	if(got_cell)
+	icon_state = "sapper[anchored?"_deployed":""][active?"_on":""]"
+	if(cell)
 		overlays += image(icon, "overlay_cell")
 
-	for(var/turf/simulated/floor/turf in range(get_turf(src), radius))
-		if(active && !turf.density)
-			if(!(turf_overlay_image in turf.overlays))
-				turf.overlays += turf_overlay_image
-		else
-			if(turf_overlay_image in turf.overlays)
-				turf.overlays -= turf_overlay_image
+	var/turf/local_turf = get_turf(src)
+	if(active)
+		for(var/turf/simulated/floor/turf in range(local_turf, radius))
+			if(bresenhem_line_check(local_turf, turf))
+				var/dx = (turf.x - local_turf.x) * 32
+				var/dy = (turf.y - local_turf.y) * 32
+				var/matrix/m = new()
+				m.Translate(dx, dy)
+				var/image/turf_overlay_image = image(icon, "overlay_turf")
+				turf_overlay_image.transform = m
+				overlays += turf_overlay_image
 
 /obj/item/weapon/floor_shocker/proc/fire(var/mob/living/mob, var/hands = FALSE)
-	if(!issilicon(mob) && power_source)
-		return electrocute_mob(mob, power_source, src, 0.4, hands)
+	if(!issilicon(mob))
+		if(powernet)
+			return electrocute_mob(mob, powernet, src, 0.4, hands)
+		if(cell)
+			return electrocute_mob(mob, cell, src, 0.4, hands)
 	return FALSE
